third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/cssom-view/elementFromPoint-003-expected.txt

This is a testharness.js-based test.
FAIL Checking whether dynamic changes to visibility interact correctly with
  table anonymous boxes assert_equals: Should hit the overlay first. expected Element node <div id="overlay"><div></div><div></div></div> but got Element node <div id="target">Some text</div>
Harness: the test ran to completion.
